test: share copy_to_tmp helper in t-data.h
copy_to_tmp was duplicated verbatim in t-circ_trott_coeff
and t-circ_trott2_coeff -- both algorithm tests need to
work on a /tmp copy of the committed fixture because
trott_simul issues per-step writes that would otherwise
mutate the on-disk reference.  Promote to t-data.h as
test_fixture_copy().

The TEST_DATA fixture array gains [[maybe_unused]] so
tests that include the header only for the helper
build clean.
