+++ test/t-circ_trott_coeff.c
@@ -11,40 +11,17 @@
 #include "ph2run/data.h"
 #include "phase2.h"
 
+#include "t-data.h"
 #include "test.h"
 
 #define WD_SEED UINT64_C(0xc91eef4d8aab2110)
 #define MARGIN (1.0e-13)
 #define STEPS (2)
 
-/* The data layer is rank-0-only and opens files RDWR, so a
- * test running on a committed fixture would mutate the file
- * via the per-step writes that trott_simul issues now.  Copy
- * the fixture to /tmp first, work on the copy, delete on the
- * way out. */
-static void copy_to_tmp(const char *src, const char *dst)
-{
-	struct world_info wd;
-	world_info(&wd);
-	if (wd.rank != 0)
-		return;
-	FILE *in = fopen(src, "rb");
-	TEST_ASSERT(in != NULL, "copy_to_tmp: open src %s", src);
-	FILE *out = fopen(dst, "wb");
-	TEST_ASSERT(out != NULL, "copy_to_tmp: open dst %s", dst);
-	char buf[4096];
-	size_t n;
-	while ((n = fread(buf, 1, sizeof buf, in)) > 0)
-		TEST_ASSERT(fwrite(buf, 1, n, out) == n,
-			"copy_to_tmp: short write to %s", dst);
-	fclose(in);
-	fclose(out);
-}
-
 static _Complex double run_trott(const char *src)
 {
 	const char *tmp = "/tmp/t-circ_trott_coeff.tmp.h5";
-	copy_to_tmp(src, tmp);
+	test_fixture_copy(src, tmp);
 
 	struct trott tt;
 	struct trott_data td = { .delta = 0.4, .steps = STEPS };
